Got HOC working properly
