+++ instagram-app/src/components/Authentication/withAuthenticate.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+
+const withAuthenticate = FirstComp =>
+class extends React.Component {
+    render() {
+        return <FirstComp />;
+    }
+}
+
+
+export default withAuthenticate;
